Decompose the executor into src/execute/, one function per file
src/execute.ts (588 lines) becomes seventeen modules under src/execute/,
each holding one function, with types.ts exporting the executor's types
and index.ts as the bucket export. The monolith test splits the same
way: each module gets a colocated suite, and the harness-driven describes
stay together in makeExecutor.test.ts.
