+++ src/execute/err.ts
@@ -1,0 +1,1 @@
+export const err = (reason: string): { ok: false; reason: string } => ({ ok: false, reason });
