+++ src/jobs/expandJobs.ts
@@ -1,6 +1,6 @@
 import { parse as parseYaml } from "yaml";
 import { UNKNOWN, type Scope, type Val } from "../expr/val.js";
-import type { JobExecutor } from "../execute.js";
+import type { JobExecutor } from "../execute/types.js";
 import { expandMatrixDetailed } from "../matrix/expandMatrixDetailed.js";
 import { jobDisplayName } from "../names/jobDisplayName.js";
 import { skippedDisplayName } from "../names/skippedDisplayName.js";
