+++ src/jobs/expandWorkflowJobs.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 import { expandWorkflowJobs } from "./expandWorkflowJobs.js";
-import type { ExecOutcome } from "../execute.js";
+import type { ExecOutcome } from "../execute/types.js";
 import type { Scope } from "../expr/val.js";
 import type { FetchWorkflow, ResolveRef, WorkflowReader, WorkflowSource } from "../types.js";
